# Aggregate trial statistics