Aggregate trial statistics to report the average of trials instead of reporting the last trial.

## After

### Ensemble Model
#### PA Output
```
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 269.4 infer/sec. Avg latency: 3705 usec (std 220 usec)
  Pass [2] throughput: 267.4 infer/sec. Avg latency: 3733 usec (std 239 usec)
  Pass [3] throughput: 268.6 infer/sec. Avg latency: 3714 usec (std 228 usec)
  Client:
    Request count: 4027
    Throughput: 268.467 infer/sec
    Avg latency: 3717 usec (standard deviation 229 usec)
    p50 latency: 3736 usec
    p90 latency: 3983 usec
    p95 latency: 4010 usec
    p99 latency: 4049 usec
    Avg HTTP time: 3699 usec (send 148 usec + response wait 3550 usec + receive 1 usec)
  Server:
    Inference count: 4809
    Execution count: 4809
    Successful request count: 4809
    Avg request latency: 3089 usec (overhead 297 usec + queue 188 usec + compute 2604 usec)

  Composing models:
  add_sub_1, version:
      Inference count: 4809
      Execution count: 4809
      Successful request count: 4809
      Avg request latency: 1505 usec (overhead 167 usec + queue 87 usec + compute input 165 usec + compute infer 797 usec + compute output 288 usec)

  add_sub_2, version:
      Inference count: 4809
      Execution count: 4809
      Successful request count: 4809
      Avg request latency: 1624 usec (overhead 170 usec + queue 101 usec + compute input 164 usec + compute infer 813 usec + compute output 375 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 268.467 infer/sec, latency 3717 usec
```

#### Verbose CSV
```
Concurrency,Inferences/Second,Client Send,Network+Server Send/Recv,Server Queue,Server Compute Input,Server Compute Infer,Server Compute Output,Client Recv,p50 latency,p90 latency,p95 latency,p99 latency,Avg latency,request/response,response wait
1,268.467,148,956,6,330,1610,664,1,3736,3983,4010,4049,3717,149,3550
```
### Sequence Model
#### PA Output

```
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using asynchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 1546 infer/sec. Avg latency: 625 usec (std 37 usec)
  Pass [2] throughput: 1525.6 infer/sec. Avg latency: 632 usec (std 40 usec)
  Pass [3] throughput: 1512.8 infer/sec. Avg latency: 636 usec (std 55 usec)
  Client:
    Request count: 22922
    Sequence count: 1144 (76.2667 seq/sec)
    Throughput: 1528.13 infer/sec
    Avg latency: 631 usec (standard deviation 45 usec)
    p50 latency: 625 usec
    p90 latency: 653 usec
    p95 latency: 694 usec
    p99 latency: 825 usec
    Avg HTTP time: 596 usec (send 34 usec + response wait 562 usec + receive 0 usec)
  Server:
    Inference count: 27542
    Execution count: 27542
    Successful request count: 27542
    Avg request latency: 277 usec (overhead 57 usec + queue 58 usec + compute input 66 usec + compute infer 74 usec + compute output 21 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 1528.13 infer/sec, latency 631 usec
```

#### Verbose CSV
```
Concurrency,Inferences/Second,Client Send,Network+Server Send/Recv,Server Queue,Server Compute Input,Server Compute Infer,Server Compute Output,Client Recv,p50 latency,p90 latency,p95 latency,p99 latency,Avg latency,request/response,response wait
1,1528.13,34,376,58,66,74,21,0,625,653,694,825,631,34,562
```
### Normal Model
#### PA Output
```
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 489.8 infer/sec. Avg latency: 2033 usec (std 206 usec)
  Pass [2] throughput: 486 infer/sec. Avg latency: 2050 usec (std 172 usec)
  Pass [3] throughput: 490.8 infer/sec. Avg latency: 2030 usec (std 200 usec)
  Client:
    Request count: 7333
    Throughput: 488.867 infer/sec
    Avg latency: 2038 usec (standard deviation 194 usec)
    p50 latency: 2090 usec
    p90 latency: 2236 usec
    p95 latency: 2279 usec
    p99 latency: 2342 usec
    Avg HTTP time: 2000 usec (send 142 usec + response wait 1857 usec + receive 1 usec)
  Server:
    Inference count: 8806
    Execution count: 8806
    Successful request count: 8806
    Avg request latency: 1395 usec (overhead 167 usec + queue 80 usec + compute input 157 usec + compute infer 785 usec + compute output 206 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 488.867 infer/sec, latency 2038 usec
```
#### Verbose CSV
```
Concurrency,Inferences/Second,Client Send,Network+Server Send/Recv,Server Queue,Server Compute Input,Server Compute Infer,Server Compute Output,Client Recv,p50 latency,p90 latency,p95 latency,p99 latency,Avg latency,request/response,response wait
1,488.867,142,664,80,157,785,206,1,2090,2236,2279,2342,2038,143,1857
```
## Before
### Ensemble Model

#### PA Output
```
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 269.8 infer/sec. Avg latency: 3697 usec (std 230 usec)
  Pass [2] throughput: 266.8 infer/sec. Avg latency: 3739 usec (std 271 usec)
  Pass [3] throughput: 269.4 infer/sec. Avg latency: 3703 usec (std 267 usec)
  Client:
    Request count: 1347
    Throughput: 269.4 infer/sec
    Avg latency: 3703 usec (standard deviation 267 usec)
    p50 latency: 3740 usec
    p90 latency: 3986 usec
    p95 latency: 4005 usec
    p99 latency: 4070 usec
    Avg HTTP time: 3671 usec (send 141 usec + response wait 3529 usec + receive 1 usec)
  Server:
    Inference count: 1616
    Execution count: 1616
    Successful request count: 1616
    Avg request latency: 3075 usec (overhead 294 usec + queue 190 usec + compute 2591 usec)

  Composing models:
  add_sub_1, version:
      Inference count: 1616
      Execution count: 1616
      Successful request count: 1616
      Avg request latency: 1492 usec (overhead 166 usec + queue 88 usec + compute input 164 usec + compute infer 787 usec + compute output 286 usec)

  add_sub_2, version:
      Inference count: 1616
      Execution count: 1616
      Successful request count: 1616
      Avg request latency: 1622 usec (overhead 167 usec + queue 102 usec + compute input 164 usec + compute infer 815 usec + compute output 374 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 269.4 infer/sec, latency 3703 usec
```
#### Verbose CSV

```
Concurrency,Inferences/Second,Client Send,Network+Server Send/Recv,Server Queue,Server Compute Input,Server Compute Infer,Server Compute Output,Client Recv,p50 latency,p90 latency,p95 latency,p99 latency,Avg latency,request/response,response wait
1,269.4,141,962,6,328,1603,660,1,3740,3986,4005,4070,3703,142,3529
```

### Sequence Model
#### PA Output
```
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using asynchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 1541.8 infer/sec. Avg latency: 626 usec (std 33 usec)
  Pass [2] throughput: 1515.2 infer/sec. Avg latency: 636 usec (std 51 usec)
  Pass [3] throughput: 1510.4 infer/sec. Avg latency: 638 usec (std 49 usec)
  Client:
    Request count: 7552
    Sequence count: 379 (75.8 seq/sec)
    Throughput: 1510.4 infer/sec
    Avg latency: 638 usec (standard deviation 49 usec)
    p50 latency: 628 usec
    p90 latency: 677 usec
    p95 latency: 710 usec
    p99 latency: 877 usec
    Avg HTTP time: 601 usec (send 36 usec + response wait 565 usec + receive 0 usec)
  Server:
    Inference count: 9071
    Execution count: 9071
    Successful request count: 9071
    Avg request latency: 278 usec (overhead 58 usec + queue 59 usec + compute input 65 usec + compute infer 74 usec + compute output 21 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 1510.4 infer/sec, latency 638 usec
```
#### Verbose CSV
```
Concurrency,Inferences/Second,Client Send,Network+Server Send/Recv,Server Queue,Server Compute Input,Server Compute Infer,Server Compute Output,Client Recv,p50 latency,p90 latency,p95 latency,p99 latency,Avg latency,request/response,response wait
1,1510.4,36,380,59,65,74,21,0,628,677,710,877,638,36,565
```

### Normal Model
#### PA Output
```
*** Measurement Settings ***
  Batch size: 1
  Using "time_windows" mode for stabilization
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 492 infer/sec. Avg latency: 2024 usec (std 212 usec)
  Pass [2] throughput: 490.8 infer/sec. Avg latency: 2030 usec (std 174 usec)
  Pass [3] throughput: 495 infer/sec. Avg latency: 2012 usec (std 212 usec)
  Client:
    Request count: 2475
    Throughput: 495 infer/sec
    Avg latency: 2012 usec (standard deviation 212 usec)
    p50 latency: 2063 usec
    p90 latency: 2221 usec
    p95 latency: 2272 usec
    p99 latency: 2332 usec
    Avg HTTP time: 1978 usec (send 141 usec + response wait 1836 usec + receive 1 usec)
  Server:
    Inference count: 2969
    Execution count: 2969
    Successful request count: 2969
    Avg request latency: 1376 usec (overhead 165 usec + queue 79 usec + compute input 156 usec + compute infer 772 usec + compute output 204 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 495 infer/sec, latency 2012 usec
```
#### Verbose CSV
```
Concurrency,Inferences/Second,Client Send,Network+Server Send/Recv,Server Queue,Server Compute Input,Server Compute Infer,Server Compute Output,Client Recv,p50 latency,p90 latency,p95 latency,p99 latency,Avg latency,request/response,response wait
1,495,141,656,79,156,772,204,1,2063,2221,2272,2332,2012,142,1836
```