@tanmayv25 2022-06-03
What is the motivation of this change? 
Instead of inference statistics of a single trial (count_window or time_window), it looks we are reporting stats from trial times  count_windows and timed_windows.? Wouldn't the latest trial run be the most stable with everything warmed up?