@Tabrizian 2022-06-03
>Instead of inference statistics of a single trial (count_window or time_window), it looks we are reporting stats from trial times count_windows and timed_windows.?

Exactly. The motivation is that we have more data from multiple trials that we are not utilizing. By combining the information from all the trials we would have better results since we have collected more samples.

> Wouldn't the latest trial run be the most stable with everything warmed up?

I think you are raising an important point. Do you think we can assume that the loaded models are warmed up using the server warm up feature?

CC @nv-braf