@tanmayv25 2022-06-03
> Do you think we can assume that the loaded models are warmed up using the server warm up feature?

Model's may be warmed up. But there are other aspects of the Inference pipeline that still needs warming up. The inference threads in the the server endpoint, the messages in the buckets(also in the endpoint). Even the perf_analyzer worker threads needs warming up. Not all worker threads would be up and running from get go. To get to a consistent concurrency it might take some time.

>  until alternative warmup mechanism is introduced in PA.

The point of trials with stability threshold was to do exactly that. The user provides us a count_window or time_window for which we must measure the statistics. The first window run is most likely the most unstable. The second window run will stabilize little bit. We use last 3(default value) trial windows to detect when the stability is achieved so that we can report the stats from apparently the most stable measurement window.

`(N-2)th window, (N-1)th,  Nth window`

If the average latency and throughput are within acceptable noise threshold then we report Nth window measurement. 
If we are reporting our throughput and latency from all these windows then I don't think measurement windows definition would hold any meaning. 

For an ideal scenario(no noise), these numbers  for all the three windows should be identical, hence it doesn't have any effect on whether the averages are reported for which one of them or all of them. But if the PA was still warming up the pipeline then, latest window is supposed to be most stable and most noise-free.