@tgerdesnv 2022-06-06
If we are determining stability based on the last N windows, then we want to include all N windows in the calculation (not ALL windows).  If N includes all windows, then we are stable enough and are ok to include what might be considered 'warmup'. Longer term there is another story to explicitly determine warmup, which will at least always exclude the first window.