@Tabrizian 2022-06-06
@tanmayv25 I chatted with Brian about the warmup issue. I think the only risk is that we could be under-reporting perf numbers for models that stabilize AND require a warm-up. Most of the warmup issues will be caught by not being able to stabilize. @nv-braf mentioned some cases that using this PR we could see much more stable results compared to what we get currently from PA.