@tanmayv25 2022-06-07
> @nv-braf mentioned some cases that using this PR we could see much more stable results compared to what we get currently from PA.

Merging the trials is equivalent to having a larger window size. And larger sample size are expected to have more stable averages. I am fine in making this change if you really need this. But we must document the change in the treatment of measurement windows as now our reported values are not for the given measurement_window but "trial * measurement window". Most of the time the user might not even care about it. But we must keep an accurate description.