@tanmayv25 2022-06-07
It's all over in this file. Some points I found:
https://github.com/triton-inference-server/client/blob/main/src/c%2B%2B/perf_analyzer/perf_analyzer.cc#L62
https://github.com/triton-inference-server/client/blob/main/src/c%2B%2B/perf_analyzer/perf_analyzer.cc#L359
https://github.com/triton-inference-server/client/blob/main/src/c%2B%2B/perf_analyzer/perf_analyzer.cc#L490
All in all, take close look into where we are talking about latencies and throughput. Also, when hitting latency threshold, should we include the averages of all three trials or should we still use the latency of the latest window.