@nv-braf 2022-06-07
Here's some data backing up my claim. This was done using a version of PA provided by Matt with his throughput fix. I ran resnet50 ten times with a BS=256, C=256, MRC=2560:

- Avg. throughput reported was 925 +/- 20 (range from 899-963)
- Mean of last 3 windows avg. throughput was 914 +/- 5 (range from 908-923)

You can clearly see how much tighter the reported throughput range is using the mean of the last 3 windows.