@Tabrizian 2022-06-08
> All in all, take close look into where we are talking about latencies and throughput. Also, when hitting latency threshold, should we include the averages of all three trials or should we still use the latency of the latest window.

Right now we are using the average.

Regarding the documentation update, I think it might be better to keep the definition of "measurement" and "trial" the same as before. I added a paragraph explaining that the numbers reported by PA are the average of the last three trials.